Done all 6

+++ Questions/q5.js
@@ -6,6 +6,7 @@ let path = require('path');
 let sPath = path.join(__dirname,"..","unorganize","abc.txt");
 let dPath = path.join(__dirname,"abc.txt");
 
+if(fs.existsSync(sPath)){
 fs.copyFileSync(sPath,dPath);
-
 fs.unlinkSync(sPath);
+}
